+++ packages/nextjs/src/index.ts
@@ -1,13 +1,86 @@
+import {
+  CookieOptions,
+  createBrowserSupabaseClient as _createBrowserSupabaseClient,
+  createServerSupabaseClient as _createServerSupabaseClient
+} from '@supabase/auth-helpers-shared';
+import {
+  GetServerSidePropsContext,
+  NextApiRequest,
+  NextApiResponse
+} from 'next';
+
 // Types
-export type { User } from '@supabase/supabase-js';
+export type { Session, User, SupabaseClient } from '@supabase/supabase-js';
 
 // Methods
-export * from './handlers';
 export * from './middleware';
-export { default as getUser } from './utils/getUser';
-export { getProviderToken } from './utils/getProviderToken';
 export { default as withPageAuth } from './utils/withPageAuth';
 export { default as withApiAuth } from './utils/withApiAuth';
-export { default as supabaseServerClient } from './utils/supabaseServerClient';
-export { supabaseClient, SupabaseClient } from './utils/initSupabase';
 export { default as logger } from './utils/log';
+
+export function createBrowserSupabaseClient<
+  Database = any,
+  SchemaName extends string & keyof Database = 'public' extends keyof Database
+    ? 'public'
+    : string & keyof Database
+>({
+  cookieOptions
+}: {
+  cookieOptions?: CookieOptions;
+} = {}) {
+  if (
+    !process.env.NEXT_PUBLIC_SUPABASE_URL ||
+    !process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
+  ) {
+    throw new Error(
+      'NEXT_PUBLIC_SUPABASE_URL and NEXT_PUBLIC_SUPABASE_ANON_KEY env variables are required!'
+    );
+  }
+
+  return _createBrowserSupabaseClient<Database, SchemaName>({
+    supabaseUrl: process.env.NEXT_PUBLIC_SUPABASE_URL,
+    supabaseKey: process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY,
+    cookieOptions
+  });
+}
+
+export function createServerSupabaseClient<
+  Database = any,
+  SchemaName extends string & keyof Database = 'public' extends keyof Database
+    ? 'public'
+    : string & keyof Database
+>(
+  context:
+    | GetServerSidePropsContext
+    | { req: NextApiRequest; res: NextApiResponse },
+  {
+    cookieOptions
+  }: {
+    cookieOptions?: CookieOptions;
+  } = {}
+) {
+  if (
+    !process.env.NEXT_PUBLIC_SUPABASE_URL ||
+    !process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY
+  ) {
+    throw new Error(
+      'NEXT_PUBLIC_SUPABASE_URL and NEXT_PUBLIC_SUPABASE_ANON_KEY env variables are required!'
+    );
+  }
+
+  return _createServerSupabaseClient<Database, SchemaName>({
+    supabaseUrl: process.env.NEXT_PUBLIC_SUPABASE_URL,
+    supabaseKey: process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY,
+    getRequestHeader: (key) => context.req.headers[key],
+    getResponseHeader: (key) => {
+      const header = context.res.getHeader(key);
+      if (typeof header === 'number') {
+        return String(header);
+      }
+
+      return header;
+    },
+    setHeader: (key, value) => context.res.setHeader(key, value),
+    cookieOptions
+  });
+}