+++ packages/react/src/index.tsx
@@ -1,5 +1,5 @@
 // Types
-export type { User } from '@supabase/supabase-js';
+export type { Session, User, SupabaseClient } from '@supabase/supabase-js';
 
 // Methods & Components
-export { UserProvider, useUser } from './components/UserProvider';
+export * from './components/SessionContext';